extract all functionality related to admin to a helper struct
